refactor: created AnalyticsCard.jsx component and extracted card generation from GroupAnalytics.jsx to there
